handled the processing of messages in a separate file
