+++ lib/index.js
@@ -17,6 +17,7 @@ module.exports = class Bot {
       hears: require('./methods/hears').bind(this),
       on: require('./methods/on').bind(this),
       listen: require('./methods/listen').bind(this),
+      message: require('./methods/message').bind(this),
       api: require('./api'),
       getLastMessage: require('./utils/getLastMessage')
     })
